feat: Integrate css modules for Card
