+++ frontend/src/components/Card/Card.module.css
@@ -1,0 +1,4 @@
+.root {
+  position: relative;
+  cursor: pointer;
+}
