+++ frontend/src/components/Card/_components/CardContent/CardContent.module.css
@@ -1,0 +1,5 @@
+.content {
+  background-color: var(--colorWhite);
+  padding: var(--spacing-sm);
+  border-radius: 0 0 var(--border-radius-sm) var(--border-radius-sm);
+}
