+++ frontend/src/components/Card/_components/CardHeader/CardHeader.module.css
@@ -1,0 +1,6 @@
+.header {
+  position: relative;
+  background-color: var(--colorDark);
+  padding: var(--spacing-md);
+  border-radius: var(--border-radius-sm) var(--border-radius-sm) 0 0;
+}
