+++ frontend/src/components/Card/_components/CardHeader/index.tsx
@@ -1,8 +1,16 @@
-import { memo } from 'react';
-import type { CardHeaderProps } from '@/components/Card/types';
+import { memo, type ReactNode } from 'react';
+import classNames from 'classnames';
+import styles from './CardHeader.module.css';
 
-export const CardHeader = memo(({ children }: CardHeaderProps) => (
-  <div className="card__header">
+export interface CardHeaderProps {
+  children: ReactNode;
+  className?: string;
+}
+
+export const CardHeader = memo(({ children, className }: CardHeaderProps) => (
+  <div className={classNames(styles.header, className)}>
     {children}
   </div>
 ));
+
+CardHeader.displayName = 'CardHeader';
