--- frontend/src/components/Card/card.css
@@ -1,17 +1,0 @@
-.card {
-  position: relative;
-  cursor: pointer;
-}
-
-.card__header {
-  position: relative;
-  background-color: var(--colorDark);
-  padding: 16px;
-  border-radius: 5px 5px 0 0;
-}
-
-.card__content {
-  background-color: var(--colorWhite);
-  padding: 8px;
-  border-radius: 0 0 5px 5px;
-}
