--- frontend/src/components/Card/types.ts
@@ -1,13 +1,0 @@
-export interface CardProps {
-  className?: string;
-  onClick?: () => void;
-  children: React.ReactNode;
-}
-
-export interface CardHeaderProps {
-  children: React.ReactNode;
-}
-
-export interface CardContentProps {
-  children: React.ReactNode;
-}
